refactor: share the paged table controls across pages
Lifts PageControls and usePageStack out of SubstratePage into
components/table, so a second paged list reuses them rather than
re-deriving the token stack. Adds a reset() for a list whose order
changes under the reader.
